+++ src/configuration.ml
@@ -30,6 +30,7 @@ end
 type t = {
   alias_barrier_modules : string list;
   constructor_map : ConstructorMapping.t list;
+  error_blacklist : string list;
   escape_value : string list;
   file_name : string;
   first_class_module_path_blacklist : string list;
@@ -46,6 +47,7 @@ type t = {
 let default (file_name : string) : t = {
   alias_barrier_modules = [];
   constructor_map = [];
+  error_blacklist = [];
   escape_value = [];
   file_name;
   first_class_module_path_blacklist = [];
@@ -69,6 +71,9 @@ let is_constructor_renamed (configuration : t) (typ : string) (name : string)
     source = name && typ' = typ
   ) |>
   Option.map (fun { ConstructorMapping.target; _ } -> target)
+
+let is_error_in_blacklist (configuration : t) (error_id : string) : bool =
+  List.mem error_id configuration.error_blacklist
 
 let is_value_to_escape (configuration : t) (name : string) : bool =
   List.mem name configuration.escape_value
@@ -177,6 +182,9 @@ let of_json (file_name : string) (json : Yojson.Basic.t) : t =
               { ConstructorMapping.source; target; typ }
             ) in
           {configuration with constructor_map = entry}
+        | "error_blacklist" ->
+          let entry = get_string_list "error_blacklist" entry in
+          {configuration with error_blacklist = entry}
         | "escape_value" ->
           let entry = get_string_list "escape_value" entry in
           {configuration with escape_value = entry}